requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==4.3.0
django-simple-history==3.4.0
Pillow==10.1.0
django-storages==1.14.2
django-ckeditor==6.7.0
wheel==0.42.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
argon2-cffi==23.1.0
PyJWT==2.8.0
python-slugify==8.0.4
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.34.54